requirements.txt
pyyaml==5.3.1
transitions==0.7.1